+++ zcash_primitives/src/transaction/components/transparent.rs
@@ -10,6 +10,7 @@ use crate::legacy::{Script, TransparentAddress};
 use super::amount::{Amount, BalanceError};
 
 pub mod builder;
+pub mod fees;
 
 pub trait Authorization: Debug {
     type ScriptSig: Debug + Clone + PartialEq;